Migrate parsley analytics to use lib methods
