+++ apps/parsley/src/analytics/loadingPage/useLogDownloadAnalytics.ts
@@ -1,4 +1,5 @@
-import { useAnalyticsRoot } from "analytics/useAnalyticsRoot";
+import { useAnalyticsRoot } from "@evg-ui/lib/analytics/hooks";
+import { AnalyticsObject } from "analytics/types";
 import { LogTypes } from "constants/enums";
 
 type Action =
@@ -22,4 +23,4 @@ type Action =
     };
 
 export const useLogDownloadAnalytics = () =>
-  useAnalyticsRoot<Action>("LoadingPage");
+  useAnalyticsRoot<Action, AnalyticsObject>("LoadingPage");
