+++ apps/parsley/src/analytics/shortcuts/useShortcutAnalytics.ts
@@ -1,5 +1,7 @@
-import { useAnalyticsRoot } from "analytics/useAnalyticsRoot";
+import { useAnalyticsRoot } from "@evg-ui/lib/analytics/hooks";
+import { AnalyticsObject } from "analytics/types";
 
 type Action = { name: "Used shortcut"; keys: string };
 
-export const useShortcutAnalytics = () => useAnalyticsRoot<Action>("Shortcut");
+export const useShortcutAnalytics = () =>
+  useAnalyticsRoot<Action, AnalyticsObject>("Shortcut");
