+++ apps/parsley/src/analytics/types.ts
@@ -1,0 +1,6 @@
+export type AnalyticsObject =
+  | "LogDrop"
+  | "LogWindow"
+  | "Preferences"
+  | "LoadingPage"
+  | "Shortcut";
