+++ packages/lib/src/utils/query-string/index.ts
@@ -1,5 +1,11 @@
 import queryString, { ParseOptions, StringifyOptions } from "query-string";
 
+/**
+ * `parseQueryString` is a function that parses a query-string string into an object
+ * @param search - The search string to parse
+ * @param options - The options to use when parsing the query-string these are the same options as the query-string library
+ * @returns - The parsed object
+ */
 export const parseQueryString = (
   search: string,
   options: ParseOptions = {},
@@ -13,6 +19,12 @@ export const parseQueryString = (
   return queryString.parse(search, parseOptions);
 };
 
+/**
+ * `stringifyQuery` is a function that stringifies an object into a query-string string
+ * @param object - The object to stringify into a query-string
+ * @param options - The options to use when stringifying the object these are the same options as the query-string library
+ * @returns - The query-string string
+ */
 export const stringifyQuery = (
   object: { [key: string]: any },
   options: StringifyOptions = {},
